rename: 컴포넌트명 변경 및 스타일코드 분리
